ci: run mobile tests separately
